+++ Modules/Whitelabels/Database/Migrations/2019_01_16_102927_create_states_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 
 class CreateStatesTable extends Migration
@@ -18,7 +19,7 @@ class CreateStatesTable extends Migration
             $table->timestamps();
         });
 
-        if (Schema::hasTable('whitelabels')) {
+        if (Schema::hasTable('whitelabels') && !Schema::hasColumn('whitelabels', 'state_id')) {
             Schema::table('whitelabels', function (Blueprint $table) {
                 $table->integer('state_id')->after('bg_image')->nullable()->unsigned()->index();
                 $table->foreign('state_id')->references('id')->on('states')->onDelete('cascade');
@@ -31,6 +32,8 @@ class CreateStatesTable extends Migration
      */
     public function down()
     {
+        DB::statement('SET FOREIGN_KEY_CHECKS = 0');
         Schema::dropIfExists('states');
+        DB::statement('SET FOREIGN_KEY_CHECKS = 1');
     }
 }
